examples - Move oneway to thrift-oneway (#644)
