--- internal/examples/oneway/.gitignore
@@ -1,1 +1,0 @@
-oneway
