+++ internal/examples/thrift-oneway/.gitignore
@@ -1,0 +1,1 @@
+thrift-oneway
