+++ internal/examples/thrift-oneway/gen.go
