+++ internal/examples/thrift-oneway/sink.thrift
