+++ internal/examples/thrift-oneway/sink/hello_sink.go
