+++ internal/examples/thrift-oneway/sink/types.go
